requirements.txt
langchain==0.2.8
langchain_google_genai==1.0.7
Pillow==10.4.0
protobuf==5.27.2
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.36.0
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.13
google-generativeai
chromadb
faiss-cpu
langchain-community


